Update code review checklist to reflect lessons learned
Now that we've reviewed a half-dozen sprints, Curtis and I have found that aspects of how we work and the how the vendor works don't align with the documented process, in ways that require modifying the checklist, rather than the process. These changes codify the reality of how we've found it's best for us to work.